+++ R/GradientBoostingMachine.R
@@ -218,13 +218,13 @@ fitXgboost <- function(
       data = dataMatrix[-trainInd, , drop = FALSE],
       label = labels$outcomeCount[-trainInd]
     )
-    watchlist <- list(train = train, test = test)
+    evals <- list(train = train, test = test)
   } else {
     train <- xgboost::xgb.DMatrix(
       data = dataMatrix,
       label = labels$outcomeCount
     )
-    watchlist <- list()
+    evals <- list()
   }
 
   outcomes <- sum(labels$outcomeCount > 0)
@@ -242,11 +242,11 @@ fitXgboost <- function(
       alpha = hyperParameters$alpha,
       objective = "binary:logistic",
       base_score = outcomeProportion,
-      eval_metric = "auc"
+      eval_metric = "auc",
+      nthread = settings$threads
     ),
-    nthread = settings$threads, # ?
     nrounds = hyperParameters$ntrees,
-    watchlist = watchlist,
+    evals = evals,
     print_every_n = 10,
     early_stopping_rounds = hyperParameters$earlyStopRound,
     maximize = TRUE